fix: Resolves issue with clicking text boxes causing pixel overflow
This commit fixes the issue with pixels overflowing when the
keyboard comes up for the user to enter input. This problem was
resolved by wrapping the top part of the login screen, and the
bottom form part of the login screen, in their own
SingleChildScrollView and ConstrainedBox widgets. The
ConstrainedBox widget allows for Spacer widgets to exist within
a scrollable view (in this case the SingleChildScrollView), which
would normally cause errors with Spacer widgets trying to take up
as much space as possible.