bug: Shows current state of app with keyboard not showing
This commit replaces TextFormField instances with a new
CustomTextFormField StatefulWidget to reduce code duplication.
Now, a bug appears in which the keyboard does not display when
the text form fields are selected. This branch is being linked
to on a Stackoverflow question in order to provide all the code
to whoever wants to look into the problem.